+++ src/Gaa.DotNet.Templates/content/gaa.project.service.grpc/ServiceLayer.cs
@@ -1,0 +1,17 @@
+using System.Diagnostics.CodeAnalysis;
+
+namespace Gaa.Project.Service.Grpc;
+
+/// <summary>
+/// Маркерный класс для сборки.
+/// </summary>
+[ExcludeFromCodeCoverage]
+public sealed class ServiceLayer
+{
+    /// <summary>
+    /// Инициализация экземпляра <see cref="ServiceLayer"/>.
+    /// </summary>
+    private ServiceLayer()
+    {
+    }
+}